Database Finished

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -24,7 +24,7 @@ class CreateUsersTable extends Migration
             $table->string('profile_photo_path', 2048)->nullable();
             $table->timestamps();
 
-            $table->integer('code');
+            $table->bigInteger('code');
         });
     }
